% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedAlphaFunctions.R
\name{ar1Solver}
\alias{ar1Solver}
\title{Compute AR1 correlation coefficient}
\usage{
ar1Solver(t)
}
\arguments{
\item{t}{A numeric vector of t-statistics}
}
\value{
Estimated AR1 correlation coefficient
}
\description{
Computes value for AR1 correlation coefficient for use in \code{p_adjust}
}
\examples{
t <- rt(1e3, df = 1)
rho <- ar1Solver(t)


}
\seealso{
\code{\link[bdots]{p_adjust}}
}
