% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsBoot.R
\name{bdotsBoot}
\alias{bdotsBoot}
\title{Create bootstrapped curves from bdotsObj}
\usage{
bdotsBoot(
  formula,
  bdObj,
  N.iter = 1000,
  alpha = 0.05,
  p.adj = "oleson",
  cores = 0,
  ...
)
}
\arguments{
\item{formula}{See details.}

\item{bdObj}{An object of class 'bdotsObj'}

\item{N.iter}{Number of iterations of bootstrap to draw}

\item{alpha}{Significance level}

\item{p.adj}{Adjustment to make to pvalues for significance. Will be able to
use anything from \code{p.adjust} function, but for now, just "oleson"}

\item{cores}{Number of cores to use in parallel. Default is zero, which
uses half of what is available.}
}
\value{
Object of class 'bdotsBootObj'
}
\description{
Creates bootstrapped curves and performs alpha adjustment. Can perform
"difference of difference" for nested comparisons
}
\details{
The formula is the only tricky part of this. There will be a minor
update to how it works in the future.  The three parts we will examine here
are Groups, the LHS, and the RHS. This is incompelete below, but is detailed
in the original vignette sent out

## Groups

To be confusing, we use "groups" in two senses here. The first use, which we
will denote with captial "G" Group, indicates the column name of the dataset
representing one of the variables by which we will distinguish categories of
observations. Lowercase "g" group will be the values in this column.

## LHS

If we are comparing two groups, we will just use the name of the outcome, i.e.,
`y ~` .
}
