% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsCorr.R
\name{bdotsCorr}
\alias{bdotsCorr}
\title{Correlation with fixed value in bdots}
\usage{
bdotsCorr(bdObj, val, ciBands = FALSE, method = "pearson")
}
\arguments{
\item{bdObj}{Object of class `bdotsObj`}

\item{val}{Character string of fixed value for correlation in dataset from `bdotsFit`}

\item{ciBands}{Boolean for including confidence intervals}

\item{method}{Arguments for `cor` or `cor.test`. The default option us `method = "pearson"`}
}
\description{
Find the correlation of a fixed value with the bdots fitted
curves at each time point
}
