% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsRefit.R
\name{bdotsRefit}
\alias{bdotsRefit}
\title{Refit Observations Returned from bdotsFit}
\usage{
bdotsRefit(bdObj, fitCode = 1, ...)
}
\arguments{
\item{bdObj}{An object of class 'bdotsObj' returned from \code{bdotsFit}}

\item{fitcode}{A length one numeric indicating observations to refit. See Details}
}
\value{
Returns bdObj with updated fit
}
\description{
Refit Observations Returned from bdotsFit
}
\details{
fitCode indicates lower bound on observations to refit. For example,
if \code{fitCode = 4}, \code{bdotsRefit} will prompt user to refit all
observations with fitCode = 4, 5, 6.
}
