% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsRefit.R
\name{bdRemove}
\alias{bdRemove}
\title{bdots Remove Function}
\usage{
bdRemove(bdObj, fitCode = 6, removePairs = TRUE)
}
\arguments{
\item{bdObj}{bdots object}

\item{fitCode}{max fitCode to retain}

\item{removePairs}{Boolean. Remove subject pairs. Default is TRUE to retain t-test}
}
\description{
Remove observations with a specified fitCode and optionally all pairs
}
\details{
Would be nice if I could give any expression and remove that instead.
But that's maybe not as useful as it sounds
}
