% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{bool2idx}
\alias{bool2idx}
\title{Need to turn  T/F into indices and I think I have an idea how}
\usage{
bool2idx(b)
}
\arguments{
\item{b}{vector of boolean}
}
\description{
Need to turn  T/F into indices and I think I have an idea how
}
