% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{brbindlist}
\alias{brbindlist}
\title{Rbindlist for bdots objects}
\usage{
brbindlist(l, ...)
}
\arguments{
\item{l}{bdotsObject}

\item{...}{for compatability with data.table::rbindlist}
}
\description{
rbindlist for bdots objects that saves attributes
}
\details{
As data.table is no longer imported, we need a separate function to prevent
masking of the generic if data.table loaded after bdots
}
