% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ci}
\alias{ci}
\title{ci dataset}
\format{
An object of class \code{data.frame} with 108216 rows and 5 columns.
}
\usage{
ci
}
\description{
ci dataset - need to include details
}
\keyword{datasets}
