% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cohort_unrelated}
\alias{cohort_unrelated}
\title{cohort_unrelated dataset}
\format{
An object of class \code{data.frame} with 50100 rows and 6 columns.
}
\usage{
cohort_unrelated
}
\description{
cohort_unrelated dataset - need to include details
}
\keyword{datasets}
