% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootHelper.R
\name{createCurveList}
\alias{createCurveList}
\title{Create old bdots curve list}
\usage{
createCurveList(x, prs, splitGroups)
}
\arguments{
\item{x}{list of group distributions}

\item{prs}{named list of parsed call}

\item{splitGroups}{splitGroups}
}
\description{
Create old bdots curve list
}
