% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{createGroupDists}
\alias{createGroupDists}
\title{Create bootstrapped distribution of groups}
\usage{
createGroupDists(x, prs, b, cores)
}
\arguments{
\item{x}{list of bdObj}

\item{prs}{list from boot parser}

\item{b}{number of bootstraps}

\item{cores}{cores}
}
\description{
Function to create bootstrapped distribution of groups depending
on paired status and difference of difference
}
