% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_cohort_unrelated}
\alias{df_cohort_unrelated}
\title{df_cohort_unrelated dataset}
\format{
An object of class \code{data.frame} with 78156 rows and 5 columns.
}
\usage{
df_cohort_unrelated
}
\description{
df_cohort_unrelated dataset - need to include details
}
\keyword{datasets}
