% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedAlphaFunctions.R
\name{effectiveAlpha_f}
\alias{effectiveAlpha_f}
\title{Effective Alpha Functional}
\usage{
effectiveAlpha_f(rho, n = 10, df = NULL, method = "norm")
}
\arguments{
\item{rho}{Correlation coefficient}

\item{n}{Number of observations}

\item{df}{Degrees of freedom if \code{method = "t"}}

\item{method}{Character string. Determines distribution for adjusted alpha
can be either \code{"norm"} for normal distribution or \code{"t"} for t-dist}
}
\description{
Functional that returns function for computing effective alpha for given
parameters and distribution
}
