% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedAlphaFunctions.R
\name{findModifiedAlpha}
\alias{findModifiedAlpha}
\title{Find modified alpha}
\usage{
findModifiedAlpha(
  rho,
  n,
  df,
  alpha = 0.05,
  errorAcc = 0.001,
  gradDiff = ifelse(cores > 3, 0.5, 0.1),
  cores = 0,
  verbose = FALSE,
  method = "t"
)
}
\arguments{
\item{rho}{correlation coefficient}

\item{n}{number of observations}

\item{df}{degrees of freedom if method == "t"}

\item{alpha}{starting alpha from which to adjust}

\item{errorAcc}{acceptable error for alphastar}

\item{gradDiff}{gradient steps in algorithm}

\item{cores}{number of cores. Default is zero, or half of what's available}

\item{verbose}{will probably remove this}

\item{method}{either "t" or "norm"}
}
\description{
find modified alpha
}
