% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveAlpha.R
\name{fwerAlpha}
\alias{fwerAlpha}
\title{fwerAlpha}
\usage{
fwerAlpha(rho, k, n = 10)
}
\arguments{
\item{rho}{Correlation coefficient}

\item{k}{Bounds of non-critical region}

\item{n}{Number of observations}
}
\description{
Family wise alpha calculation
}
\details{
Returns effective alpha, given number of tests and the correlation
coefficient. This isn't explicitly checked, but there is no reason this function
should take any non-scalar values. Derivation of this can be found on pg 12
of Jake's 'Detecting time-specific differences'. This function performs the
expression \deqn{1 - P(I_t)P(I_t \ | \ I_{t-1})^{N-1}}
}
