% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{getBootDistPaired}
\alias{getBootDistPaired}
\title{Create distribution for paired groups}
\usage{
getBootDistPaired(cl, x, b)
}
\arguments{
\item{cl}{cluster for parallel}

\item{x}{A list of bdObj}

\item{b}{Number of bootstraps}
}
\description{
Create distribution for paired groups
}
