% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{getBootDistUnpaired}
\alias{getBootDistUnpaired}
\title{Create Distribution for Groups}
\usage{
getBootDistUnpaired(x, b = 1000)
}
\arguments{
\item{x}{A subset object from bdObj}

\item{b}{Number of bootstraps}
}
\description{
Create Distribution for Groups
}
