% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsCorr.R
\name{getFitCorforGroups}
\alias{getFitCorforGroups}
\title{Get Fit Correlations}
\usage{
getFitCorforGroups(x, val, ciBands = FALSE, method = "pearson")
}
\arguments{
\item{x}{A split object of class `bdObj` split by identifiers}

\item{val}{Fixed value from dataset}

\item{ciBands}{boolean for including cibands}

\item{method}{method for correlation function}
}
\description{
Helper function for finding correlation of fixed value and
fitted values within group
}
