% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getSubCurveValues}
\alias{getSubCurveValues}
\title{Return fitted values}
\usage{
getSubCurveValues(bd, origNames = TRUE, origTime = TRUE)
}
\arguments{
\item{bd}{Single row of bdObj}

\item{origNames}{use original names for y and time, or use "y" and "time"}

\item{origTime}{Boolean. Do I actually want fitted values at observed times for that
subject, or data.table with fitted values at the union of times}
}
\description{
Returns fitted values at observed times
}
\details{
Given a single row of bdObj, this returns fitted values at the observed
times to use in conjunction with whatever else
}
