% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{getT}
\alias{getT}
\title{Generate t-distribution for permutation}
\usage{
getT(x, idx, group, whole = FALSE)
}
\arguments{
\item{x}{bdObj}

\item{idx}{permutation to use}

\item{group}{group that we are permuting against}

\item{whole}{return vector of T stats or just the max}
}
\description{
Creates vector of t-statistics for given permutation returning
either entire vector or max values
}
