% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{getTDist}
\alias{getTDist}
\title{Get null distribution of permutation}
\usage{
getTDist(x, P = 1000)
}
\arguments{
\item{x}{bdObj}

\item{P}{number of permutations}
}
\description{
Get null distribution of permutation
}
