% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{isDODpaired}
\alias{isDODpaired}
\title{Difference of difference pairing}
\usage{
isDODpaired(x, prs)
}
\arguments{
\item{x}{list of bdObjs}

\item{prs}{parsed boot objects}
}
\description{
Took the easy, inefficient way for now
}
\details{
Function to determine if list of 4 groups indicating difference of
difference is paired or not
}
