% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFunctions.R
\name{linear}
\alias{linear}
\title{Linear curve function}
\usage{
linear(dat, y, time, params = NULL, ...)
}
\arguments{
\item{dat}{subject data to be used}

\item{y}{outcome variable}

\item{time}{time variable}

\item{params}{\code{NULL} unless user wants to specify starting parameters for gnls}

\item{...}{just in case}
}
\description{
Linear function used in fitting nlme curve for observations
}
\details{
Don't use this function please

\code{y ~ slope*time + intercept}
}
