% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic curve function for nlme}
\usage{
logistic(dat, y, time, params = NULL, ...)
}
\description{
Logistic function used in fitting nlme curve for observations
}
\details{
\code{y ~ mini + (peak - mini) / (1 + exp(4 * slope * (cross - (time)) / (peak - mini)))}
}
