% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFunctions.R
\name{logistic_sac}
\alias{logistic_sac}
\title{Logistic saccade curve function for nlme}
\usage{
logistic_sac(dat, y, time, params = NULL, startSamp = 8, ...)
}
\arguments{
\item{dat}{subject data to be used}

\item{y}{outcome variable}

\item{time}{time variable}

\item{params}{\code{NULL} unless user wants to specify starting parameters for gnls}

\item{startSamp}{how many samples from distribution should we use to investigate}

\item{...}{just in case}
}
\description{
And even cooler logistic fitting function, will think of a better name later
}
\details{
\code{y ~ mini + (peak - mini) / (1 + exp(4 * slope * (cross - (time)) / (peak - mini)))}
}
