% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padjust.R
\name{p_adjust}
\alias{p_adjust}
\title{Adjust P-values for Multiple Comparisons}
\usage{
p_adjust(p, method = "oleson", n = length(p), alpha = 0.05, df, rho, cores = 0)
}
\arguments{
\item{p}{numeric vector of p-values (possibly with NAs).}

\item{method}{correction method, a character string. Can be any of the methods in
p.adjust.methods, with the additional value \code{method = "oleson"}}

\item{n}{number of comparisons, must be at least \code{length(p)}; only set this
(to non-default) when you know what you are doing!}

\item{alpha}{adjustment to be made with method oleson}

\item{df}{degrees of freedom, if using \code{method = "oleson"}}

\item{rho}{correlation coefficient, if using \code{method = "oleson"}}

\item{cores}{number of cores for use in parallel, only valid for
\code{method = "oleson"}. Default is zero, using half of the available cores}
}
\description{
Identical to stats::p.adjust, but includes \code{method = "oleson"}
}
\details{
I will revist this function later to make sure the arguments are consistent
with how they may be used in the wild. Also need to include details from p.adjust in stats
I can add oleson there with there references. Need example too, I think. ALSO - it's sometimes
helpful to know alphastar. Should I return a similar value for the other methods? Do they have
an equivalent? I know BH does, but the rest are not linear adjustments.
ALSO::: can ar1Solver determine rho from pvalue, or is tvalue needed?
}
