% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permTest}
\alias{permTest}
\title{Permutation testing for bboot}
\usage{
permTest(x, prs, alpha, P, cores = detectCores() - 1L)
}
\arguments{
\item{x}{this is splitGroups (each item of list bdotsObj)}

\item{prs}{from parser, indicates inner/outer groups}

\item{alpha}{alpha for fwer}

\item{P}{number of permutations}
}
\description{
Provides alternative for controlling FWER
}
