% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.bdotsBootObj}
\alias{plot.bdotsBootObj}
\title{Plot for object of class bdotsBootObj}
\usage{
\method{plot}{bdotsBootObj}(x, alpha = NULL, ciBands = TRUE, plotDiffs = TRUE, group = NULL, ...)
}
\arguments{
\item{x}{An object of class bdotsBootObj}

\item{alpha}{Significance level for plotting confidence intervals.}

\item{ciBands}{Boolean indicating whether or not to include confidence intervals
around fitted curves}

\item{plotDiffs}{Boolean to plot difference curve}

\item{group}{Specify group to plot if difference of difference was used. The
user can also subset the bdotsBootObj prior to plotting. Currently not used}

\item{...}{ignore for now, but will eventually allow plot parameters}
}
\value{
List of ggplot objects, which may be helpful if the margins are weird
}
\description{
Allows a number of different but also unstable option for plotting an object
of class bdotsBoot
}
\details{
This plot function is also a bit unstable and is expected to change
}
