% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsCorr.R
\name{plot.bdotsCorrObj}
\alias{plot.bdotsCorrObj}
\title{Plots for bdotsCorr}
\usage{
\method{plot}{bdotsCorrObj}(x, ciBands = FALSE, window = NULL, ...)
}
\arguments{
\item{x}{object of class `bdotsCorrObj`}

\item{ciBands}{boolean. Whether or not to include confidence intervals in plots.
Must have been selected in `bdotsCorr`}

\item{window}{A length 2 numeric vector with start and end points
for the  plotting window}

\item{...}{super secret, don't use}
}
\description{
Plots correlation of fixed value wtih fitted curves over time
}
