% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bdotsBootObj}
\alias{print.bdotsBootObj}
\title{Print `bdotsBootObj`}
\usage{
\method{print}{bdotsBootObj}(x, ...)
}
\arguments{
\item{x}{An object of class `bdotsBootObj`}

\item{...}{Top secret alpha one code red}
}
\description{
Prints argument. Really, just the summary function
}
\details{
Generic for printing `bdotsBootObj`
}
