% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{print.bdotsBootSummary}
\alias{print.bdotsBootSummary}
\title{Print bdotsBoot Summary}
\usage{
\method{print}{bdotsBootSummary}(x, ...)
}
\arguments{
\item{x}{generic name, but this will be an object of bdotsBootSummary}

\item{...}{ignored for now}
}
\description{
That's pretty much it. This is a print method, so there is likely
not much need to call it directly
}
