% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_ttest.R
\name{print.bdotsPars_ttest2}
\alias{print.bdotsPars_ttest2}
\title{Print Parameter Test Summary}
\usage{
\method{print}{bdotsPars_ttest2}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{not used}
}
\description{
Print Parameter Test Summary
}
\details{
That's pretty much it. This is a print method, so there is likely
not much need to call it directly
}
