% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{print.bdotsSummary}
\alias{print.bdotsSummary}
\title{Print bdotsObj Summary}
\usage{
\method{print}{bdotsSummary}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{not used}
}
\description{
Stuff
}
\details{
That's pretty much it. This is a print method, so there is likely
not much need to call it directly
}
