% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rbindlist.bdObjList}
\alias{rbindlist.bdObjList}
\title{rbindlist for bdotsObjects}
\usage{
\method{rbindlist}{bdObjList}(x, ...)
}
\arguments{
\item{x}{bdotsObject}

\item{...}{for compatability with data.table}
}
\description{
Similar to data.table::rbindlist, but preserves botsObjects attributes
}
