% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{split.bdotsObj}
\alias{split.bdotsObj}
\title{Split object of class bdotsObj}
\usage{
\method{split}{bdotsObj}(x, f, drop = FALSE, by, ...)
}
\arguments{
\item{x}{Object of class bdotsObj}

\item{f}{For consistency with generic, but is not used}

\item{drop}{logical. Default FALSE will not drop empty list elements caused
by factor levels not referred by that factor. Analagous to data.table::split}

\item{by}{Character vector of column names on which to split. Usually will
be Subject or one of the fitted groups}

\item{...}{not used}
}
\description{
Analogous to other splitting functions, but retains necessary attributes
across the split object. As of now, it can only be unsplit with bdots::rbindlist
}
