% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset.bdotsBootObj}
\alias{subset.bdotsBootObj}
\title{Subset a nested group bdotsBoot objects}
\usage{
\method{subset}{bdotsBootObj}(x, group, adjustAlpha = NULL, ...)
}
\arguments{
\item{x}{An object returned from \code{bdotsBoot}}

\item{group}{A group to subset. Must be an outer group}

\item{adjustAlpha}{currently not used. Will give option to recompute adjusted alpha}

\item{...}{Not used}
}
\description{
Subset a nested group bdotsBoot objects
}
\details{
This function is used to subset a bdotsBootObject that was fit to compute
the difference of differences. This allows the user to subset out the outer group
in the comparison for plotting and investigation
}
