% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary.bdotsBootObj}
\alias{summary.bdotsBootObj}
\title{Summary for bdotsBootObj}
\usage{
\method{summary}{bdotsBootObj}(object, ...)
}
\arguments{
\item{object}{An object of class bdotsObj}

\item{...}{Ignored for now}
}
\value{
Returns an object of class "bdotsBootSummary". There is some summarized
information included if assigned to an object, i.e., `summ <- summary(bdBootObj)`
then `str(summ)`
}
\description{
Provides summary information for bdotsBootObj
}
