% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{target}
\alias{target}
\title{target dataset}
\format{
An object of class \code{data.frame} with 25050 rows and 4 columns.
}
\usage{
target
}
\description{
target dataset - need to include details
}
\keyword{datasets}
