% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{bdotsBoot}
\alias{bdotsBoot}
\title{Create bootstrapped curves from bdotsObj}
\usage{
bdotsBoot(
  formula,
  bdObj,
  Niter = 1000,
  alpha = 0.05,
  padj = "oleson",
  permutation = FALSE,
  skipDist = FALSE,
  singleMeans = FALSE,
  permAddVar = TRUE,
  cores = 0,
  ...
)
}
\arguments{
\item{formula}{See details.}

\item{bdObj}{An object of class 'bdotsObj'}

\item{Niter}{Number of iterations of bootstrap to draw}

\item{alpha}{Significance level}

\item{padj}{Adjustment to make to pvalues for significance. Will be able to
use anything from \code{p.adjust} function, but for now, just "oleson"}

\item{permutation}{Boolean indicating whether to use permutation testing rather
thank adjusting alpha to control FWER. WARNING: This option is very much in beta testing
and not recommended for general use. Also not available for paired tests or difference of difference}

\item{skipDist}{do not use}

\item{singleMeans}{definitely do not use}

\item{permAddVar}{Boolean. Add observed variability for perms?}

\item{cores}{Number of cores to use in parallel. Default is zero, which
uses half of what is available.}

\item{...}{not used}
}
\description{
DEPRECATED. Use \code{bboot} instead
}
\details{
Deprecated. Use \code{bboot} instead
}
