% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsFitter.R
\name{bdotsFitter}
\alias{bdotsFitter}
\title{Fits Individual Subject Curve}
\usage{
bdotsFitter(
  dat,
  curveType,
  rho,
  numRefits = 0,
  verbose,
  getCovOnly = NULL,
  params = NULL,
  splitVars = NULL,
  datVarNames = NULL,
  ...
)
}
\arguments{
\item{dat}{data for single subject/group combo}

\item{curveType}{this is actually a function. Should rename}

\item{rho}{correlation coefficient}

\item{numRefits}{number of refit attempts}

\item{verbose}{not used}

\item{getCovOnly}{only find covariance matrix from starting parameter values}

\item{params}{starting parameters, if wanting to add manually}

\item{splitVars}{variables used to identify group. Might combine with datVarNames}

\item{datVarNames}{character vector indicating reponse and time values from parent call}

\item{...}{not used}
}
\description{
The one subject version of bdotsFit
}
