% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsFitter.R
\name{bdotsFitter}
\alias{bdotsFitter}
\title{Fits Individual Subject Curve}
\usage{
bdotsFitter(
  dat,
  curveType,
  rho,
  numRefits = 0,
  verbose,
  thenames = NULL,
  get.cov.only = NULL,
  params = NULL,
  splitVars = NULL,
  datVarNames = NULL,
  ...
)
}
\description{
The one subject version of bdotsFit
}
