% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsRefit.R
\name{bdRemove}
\alias{bdRemove}
\title{bdots Remove Function}
\usage{
bdRemove(bdObj, fitCode = 6L, removePairs = TRUE)
}
\arguments{
\item{bdObj}{bdots object}

\item{fitCode}{min fitCode to remove. Default is 6, which removes all subjects with NULL fits (fitCode = 5 would remove 5 and 6)}

\item{removePairs}{Boolean. Remove subject pairs is one of pair is removed.
Default is TRUE to retain paired t-test}
}
\description{
Remove observations with a specified fitCode and optionally all pairs
}
\details{
This function is used to remove all bdots observations with a fit code
equal to or larger than the argument passed to \code{fitCode} without refitting.
If \code{removePairs = TRUE}, all entries for a subject will be removed if their
fit failed in any of the groups in which they were a member
}
