\name{ci}
\docType{data}
\alias{ci}

\title{ Eyetracking Data from Normal Hearing Individuals and those with Cochlear Implants}

\description{ This data set has stuff on CIs and NHs. Provide more info here.}

\usage{ci}

\format{A matrix of values}

\references{Farris-Trimble, A., McMurray, B., Cigrand, N. and Tomblin, J.B. (2014)
	The process of spoken word recognition in the face of signal degradation: Cochlear
	implant users and normal-hearing listeners. Journal of Experimental
	Psychology: Human Perception and Performance, 40(1), 308-327}

\keyword{datasets}