% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coef.bdotsObj}
\alias{coef.bdotsObj}
\title{Extract bdotsFit Moedel Coefficients}
\usage{
\method{coef}{bdotsObj}(object, ...)
}
\arguments{
\item{object}{A bdotsObj}

\item{...}{not used}
}
\description{
Returns coefficient matrix for bdotsFit object
}
