% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdotsFitter.R
\name{curveFitter}
\alias{curveFitter}
\title{Curve Fitter}
\usage{
curveFitter(dat, ff, params, rho, numRefits = 0, getCovOnly = NULL, ...)
}
\arguments{
\item{dat}{data used in building curve}

\item{ff}{formula used in buildilng curve}

\item{params}{starting parameters}

\item{rho}{correlation coefficient}

\item{numRefits}{number of refit attempts}

\item{getCovOnly}{only find covariance matrix from starting parameter values}

\item{...}{don't know that this is used, can maybe get rid of it}
}
\description{
Used in bdotsFit
}
