% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFunctions.R
\name{expCurve}
\alias{expCurve}
\title{Exponential curve function}
\usage{
expCurve(dat, y, time, params = NULL, ...)
}
\arguments{
\item{dat}{subject data to be used}

\item{y}{outcome variable}

\item{time}{time variable}

\item{params}{\code{NULL} unless user wants to specify starting parameters for gnls}

\item{...}{just in case}
}
\description{
Exponential function used in fitting nlme curve for observations
}
\details{
Remove any values of zero, or jitter, before using with bdotsFit

\code{y ~ x_0 exp(k beta)}
}
