% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModifiedAlpha.R
\name{findModifiedAlpha}
\alias{findModifiedAlpha}
\title{Find modified alpha}
\usage{
findModifiedAlpha(
  rho,
  n,
  df,
  alpha = 0.05,
  errorAcc = 0.001,
  gradDiff = ifelse(cores > 3, 0.5, 0.1),
  cores = 0,
  verbose = FALSE,
  method = "t"
)
}
\description{
find modified alpha
}
