% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFunctions.R
\name{logistic}
\alias{logistic}
\title{Logistic curve function for nlme}
\usage{
logistic(dat, y, time, params = NULL, ...)
}
\arguments{
\item{dat}{subject data to be used}

\item{y}{outcome variable}

\item{time}{time variable}

\item{params}{\code{NULL} unless user wants to specify starting parameters for gnls}

\item{...}{just in case}
}
\description{
Logistic function used in fitting nlme curve for observations
}
\details{
\code{y ~ mini + (peak - mini) / (1 + exp(4 * slope * (cross - (time)) / (peak - mini)))}
}
