% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.bdotsBootObj}
\alias{plot.bdotsBootObj}
\title{Plot for object of class bdotsBootObj}
\usage{
\method{plot}{bdotsBootObj}(x, alpha = 0.05, diffs = NULL, group = NULL, ...)
}
\arguments{
\item{x}{An object of class bdotsBootObj}

\item{alpha}{Significance level for plotting shaded regions}

\item{diffs}{Boolean to plot "difference of difference" curve}

\item{group}{Specify group to plot if difference of difference was used}

\item{...}{ignore for now}
}
\description{
Allows a number of different but also unstable option for plotting an object
of class bdotsBoot
}
\details{
Use with care
}
