% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary.bdotsBootObj}
\alias{summary.bdotsBootObj}
\title{Summary for bdotsBootObj}
\usage{
\method{summary}{bdotsBootObj}(bdBootObj, ...)
}
\arguments{
\item{bdBootObj}{An object of class bdotsObj}

\item{Ignored}{for now}
}
\description{
Provides summary information for bdotsBootObj
}
