% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary.bdotsObj}
\alias{summary.bdotsObj}
\title{Summary for bdotsObj}
\usage{
\method{summary}{bdotsObj}(bdObj, ...)
}
\arguments{
\item{bdObj}{An object of class bdotsObj}
}
\value{
Returns an object of class "bdotsSummary". There is some summarized
information included if assigned to an object, i.e., `summ <- summary(bdObj)`
then `str(summ)`
}
\description{
Provides summary information for bdotsObj
}
