% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary.bdotsObj}
\alias{summary.bdotsObj}
\title{Summary for bdotsObj}
\usage{
\method{summary}{bdotsObj}(bdObj, ...)
}
\arguments{
\item{bdObj}{An object of class bdotsObj}
}
\description{
Provides summary information for bdotsObj
}
