% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSV.R
\name{writeCSV}
\alias{writeCSV}
\title{Write fits from \code{bdotsBoot} to csv file}
\usage{
writeCSV(bootObj, file, alpha = 0.05, ...)
}
\arguments{
\item{bootObj}{An object of class \code{bdotsBootObj}}

\item{file}{file name to write out csv}

\item{alpha}{alpha level for upper/lower CI}

\item{...}{Other arguments passed to \code{data.table::fread}}
}
\description{
The function is used to write out columns for each group for which a curve
was bootstrapped
}
\details{
This is potentially useful for constructing plots in a separate
application. There is an additional column, \code{Significant} indicating if
a particular time point was considered significant between the difference curves.
For difference of difference objects, this only indicates significance for the
outer difference.
}
